Proper filename for TestTextField in line with TestWidgetsApp etc.
