+++ packages/flutter/test/widgets/editable_text_cursor_test.dart
@@ -16,6 +16,7 @@ import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 
+import 'editable_text_tester.dart';
 import 'editable_text_utils.dart';
 
 const TextStyle textStyle = TextStyle();
