+++ packages/flutter/test/widgets/routes_test.dart
@@ -11,7 +11,7 @@ import 'package:flutter/physics.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'editable_text_utils.dart' show TestTextField;
+import 'editable_text_tester.dart' show TestTextField;
 import 'semantics_tester.dart';
 
 final List<String> results = <String>[];
